PM: DefaultSigVerifier
We want to work with this simple implementation for now because we
don't want to interact with the Broker contract every time we validate
a sig. Once we have some solution for caching approved signers from the
Broker contract, we will switch over to that verifier implementation.
