+++ eth/helpers.go
@@ -6,7 +6,6 @@ import (
 	"time"
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/golang/glog"
 	"github.com/livepeer/go-livepeer/common"
 )
@@ -15,18 +14,6 @@ var (
 	BlocksUntilFirstClaimDeadline = big.NewInt(200)
 	maxUint256                    = new(big.Int).Sub(new(big.Int).Exp(big.NewInt(2), big.NewInt(256), nil), big.NewInt(1))
 )
-
-func VerifySig(addr ethcommon.Address, msg, sig []byte) bool {
-	personalMsg := fmt.Sprintf("\x19Ethereum Signed Message:\n%d%s", 32, msg)
-	personalHash := crypto.Keccak256([]byte(personalMsg))
-
-	pubkey, err := crypto.SigToPub(personalHash, sig)
-	if err != nil {
-		return false
-	}
-
-	return crypto.PubkeyToAddress(*pubkey) == addr
-}
 
 func FormatUnits(baseAmount *big.Int, name string) string {
 	amount := FromBaseUnit(baseAmount)
