+++ pm/helpers.go
@@ -2,10 +2,25 @@ package pm
 
 import (
 	"crypto/rand"
+	"fmt"
 	"testing"
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
 )
+
+// TODO refactor to a package that both eth and pm can import
+func VerifySig(addr ethcommon.Address, msg, sig []byte) bool {
+	personalMsg := fmt.Sprintf("\x19Ethereum Signed Message:\n%d%s", 32, msg)
+	personalHash := crypto.Keccak256([]byte(personalMsg))
+
+	pubkey, err := crypto.SigToPub(personalHash, sig)
+	if err != nil {
+		return false
+	}
+
+	return crypto.PubkeyToAddress(*pubkey) == addr
+}
 
 func RandHashOrFatal(t *testing.T) ethcommon.Hash {
 	key, err := randBytes(32)
